[CP-3525] Reduced boilerplate by removing duplicate adapter layer
