+++ libs/app-mtp/src/lib/app-mtp.factory.ts
@@ -6,13 +6,14 @@
 import { MtpInterface } from "./app-mtp.interface"
 import { DotnetMtp } from "./dotnet-mtp/dotnet-mtp"
 import { NodeMtp } from "./node-mtp/node-mtp"
+import { NodeMtpDeviceManager } from "Libs/app-mtp/src/lib/node-mtp/node-mtp-device-manager"
 
 export class MtpFactory {
   static createInstance(): MtpInterface {
     if (process.platform === "win32") {
       return new DotnetMtp()
     } else {
-      return new NodeMtp()
+      return new NodeMtp(new NodeMtpDeviceManager())
     }
   }
 }
